requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-csp==3.5
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.2
pipenv==2018.11.26
psycopg2==2.8.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
